Conclusão de visibilidade dos botoes

+++ src/main/java/com/projeto/view/PanelRelatorioDeUsuario.java
@@ -3,6 +3,7 @@ package com.projeto.view;
 import javax.swing.JPanel;
 import javax.swing.DefaultComboBoxModel;
 import javax.swing.GroupLayout;
+import javax.swing.ImageIcon;
 import javax.swing.GroupLayout.Alignment;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
@@ -11,9 +12,12 @@ import java.awt.Font;
 import javax.swing.JTextField;
 import javax.swing.LayoutStyle.ComponentPlacement;
 import javax.swing.JComboBox;
+import javax.swing.JFileChooser;
 import javax.swing.JButton;
 import javax.swing.JTable;
 import java.awt.event.ActionListener;
+import java.awt.event.FocusEvent;
+import java.awt.event.FocusListener;
 import java.util.ArrayList;
 import java.util.List;
 import java.awt.event.ActionEvent;
@@ -22,6 +26,7 @@ import javax.swing.table.DefaultTableModel;
 
 import com.projeto.controller.UsuarioController;
 import com.projeto.enums.TipoUsuarioEnum;
+import com.projeto.enums.TurnoEnum;
 import com.projeto.model.entity.AlunoVO;
 import com.projeto.model.entity.CoordenadorVO;
 import com.projeto.model.entity.ProfessorVO;
@@ -29,6 +34,10 @@ import com.projeto.model.entity.UsuarioVO;
 import com.projeto.seletor.RelatorioDeUsuarioSeletor;
 
 import com.projeto.repository.Constants;
+import com.projeto.repository.GeradorPlanilhaUsuario;
+
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
 
 public class PanelRelatorioDeUsuario extends JPanel {
 	private static final int TAMANHO_PAGINA = 10;
@@ -46,6 +55,13 @@ public class PanelRelatorioDeUsuario extends JPanel {
 	private String[] nomesColunas = {"Nome", "Tipo de Usuario", "Turno", "Sexo", "Possui Deficiência", "RG", "CPF"};
 	private DefaultTableModel model;
 	private int paginaAtual = 1;
+	private JButton btnGerarXls;
+	private JComboBox cbxTurno;
+	private UsuarioController usuarioController = new UsuarioController();
+	private JButton btnExcluir;
+	private JButton btnAlterar;
+	private int paginasTotal;
+	private JLabel lblTotalPaginas;
 
 	/** TIPO DE USUARIO, NOME COM LIKE, PESQUISAR TODOS. LEMBRAR DE ACRESCENTAR PAGINAÇÃO e relatório excel.
 	 * Create the panel.
@@ -55,8 +71,8 @@ public class PanelRelatorioDeUsuario extends JPanel {
 		panel = new JPanel();
 		GroupLayout groupLayout = new GroupLayout(this);
 		groupLayout.setHorizontalGroup(
-			groupLayout.createParallelGroup(Alignment.LEADING)
-				.addGroup(groupLayout.createSequentialGroup()
+			groupLayout.createParallelGroup(Alignment.TRAILING)
+				.addGroup(Alignment.LEADING, groupLayout.createSequentialGroup()
 					.addContainerGap()
 					.addComponent(panel, GroupLayout.PREFERRED_SIZE, 780, Short.MAX_VALUE)
 					.addContainerGap())
@@ -65,7 +81,7 @@ public class PanelRelatorioDeUsuario extends JPanel {
 			groupLayout.createParallelGroup(Alignment.LEADING)
 				.addGroup(groupLayout.createSequentialGroup()
 					.addContainerGap()
-					.addComponent(panel, GroupLayout.DEFAULT_SIZE, 756, Short.MAX_VALUE)
+					.addComponent(panel, GroupLayout.DEFAULT_SIZE, 712, Short.MAX_VALUE)
 					.addContainerGap())
 		);
 		
@@ -80,9 +96,12 @@ public class PanelRelatorioDeUsuario extends JPanel {
 		
 		JLabel lblTipoDeUsuario = new JLabel("Tipo de Usuário:");
 		
+
+		ArrayList<String> tipoUsuario = usuarioController.consultarTipoUsuarioController();
 		cbxTipoUsuario = new JComboBox();
-		cbxTipoUsuario.setModel(new DefaultComboBoxModel(new String[] {"TODOS", "ALUNO", "PROFESSOR", "COORDENADOR"}));
-		
+		DefaultComboBoxModel preencherTipoUsuario = new DefaultComboBoxModel(tipoUsuario.toArray());
+		cbxTipoUsuario.setModel(preencherTipoUsuario);
+
 		btnBuscar = new JButton("Buscar");
 		btnBuscar.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
@@ -97,6 +116,14 @@ public class PanelRelatorioDeUsuario extends JPanel {
 				limparTabelaUsuario();
 				txtNome.setText("");
 				cbxTipoUsuario.setSelectedIndex(0);
+				cbxTurno.setSelectedIndex(0);
+				btnExcluir.setEnabled(false);
+				btnGerarXls.setEnabled(false);
+				btnAlterar.setEnabled(false);
+				paginaAtual = 1;
+				paginasTotal = 1;
+				lblPaginaAtual.setText(paginaAtual + "");
+				lblTotalPaginas.setText(paginasTotal + "");
 			}
 		});
 		btnLimpar.setFont(new Font("Tahoma", Font.BOLD, 14));
@@ -122,6 +149,27 @@ public class PanelRelatorioDeUsuario extends JPanel {
 				return columnEditables[column];
 			}
 		});
+		tblListaDeUsuarios.addFocusListener(new FocusListener() {
+			
+			@Override
+			public void focusLost(FocusEvent e) {
+				if (tblListaDeUsuarios.getSelectedRow() > 0) {
+					btnExcluir.setEnabled(true);
+					btnAlterar.setEnabled(true);
+				} else {
+					btnExcluir.setEnabled(false);
+					btnAlterar.setEnabled(false);
+				}
+			}
+			
+			@Override
+			public void focusGained(FocusEvent e) {
+				if (tblListaDeUsuarios.getSelectedRow() >= 0) {
+					btnExcluir.setEnabled(true);
+					btnAlterar.setEnabled(true);
+				}
+			}
+		});
 		
 		btnPaginaAnterior = new JButton("← ");
 		btnPaginaAnterior.addActionListener(new ActionListener() {
@@ -133,6 +181,7 @@ public class PanelRelatorioDeUsuario extends JPanel {
 			}
 		});
 		btnPaginaAnterior.setAlignmentX(Component.CENTER_ALIGNMENT);
+		btnPaginaAnterior.setEnabled(false);
 		
 		lblPaginaAtual = new JLabel("");
 		lblPaginaAtual.setText(paginaAtual + "");
@@ -146,28 +195,37 @@ public class PanelRelatorioDeUsuario extends JPanel {
 				}
 			}
 		);
+		btnProximaPagina.setEnabled(false);
 		
-		JButton btnExcluir = new JButton("Excluir");
+		verificarBotoesPaginas();
+		
+		btnExcluir = new JButton("Excluir");
 		btnExcluir.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent arg0) {
+				
 				UsuarioController usuarioController = new UsuarioController();
 				int indiceSelecionadoNaTablela = tblListaDeUsuarios.getSelectedRow();
-				UsuarioVO usuarioSelecionado = usuario.get(indiceSelecionadoNaTablela - 1);
-				
-				String perguntaExclusao = "Deseja excluir o usuario: " + usuarioSelecionado.getNome() + "?";
-				
-				int opcaoSelecionada = JOptionPane.showConfirmDialog(null, perguntaExclusao, "Está certo disso?", JOptionPane.YES_NO_OPTION);
-				
-				if (opcaoSelecionada == JOptionPane.YES_OPTION) {
-					String mensagem = usuarioController.excluirUsuarioController(usuarioSelecionado.getIdUsuario());
-					JOptionPane.showMessageDialog(null, mensagem);
-					limparTabelaUsuario();
+				if (indiceSelecionadoNaTablela > 0) {
+					UsuarioVO usuarioSelecionado = usuario.get(indiceSelecionadoNaTablela - 1);
+					
+					String perguntaExclusao = "Deseja excluir o usuario: " + usuarioSelecionado.getNome() + "?";
+					
+					int opcaoSelecionada = JOptionPane.showConfirmDialog(null, perguntaExclusao, "Está certo disso?", JOptionPane.YES_NO_OPTION);
+					
+					if (opcaoSelecionada == JOptionPane.YES_OPTION) {
+						String mensagem = usuarioController.excluirUsuarioController(usuarioSelecionado.getIdUsuario());
+						JOptionPane.showMessageDialog(null, mensagem);
+						limparTabelaUsuario();
+					}
+				} else {
+					btnExcluir.setEnabled(false);
 				}
 			}
 		});
 		btnExcluir.setFont(new Font("Tahoma", Font.BOLD, 14));
+		btnExcluir.setEnabled(false);
 		
-		JButton btnAlterar = new JButton("Alterar");
+		btnAlterar = new JButton("Alterar");
 		btnAlterar.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				// TODO fazer alteração do usuario e chamar o painel de cadastro passando o usuario como parametro
@@ -175,46 +233,102 @@ public class PanelRelatorioDeUsuario extends JPanel {
 			}
 		});
 		btnAlterar.setFont(new Font("Tahoma", Font.BOLD, 14));
+		btnAlterar.setEnabled(false);
+		
+		btnGerarXls = new JButton("Salvar Excel");
+		btnGerarXls.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				JFileChooser jfc = new JFileChooser();
+				jfc.setDialogTitle("Salvar relatório como...");
+
+				int resultado = jfc.showSaveDialog(null);
+				if (resultado == JFileChooser.APPROVE_OPTION) {
+					String caminhoEscolhido = jfc.getSelectedFile().getAbsolutePath();
+					GeradorPlanilhaUsuario geradorPlanilha = new GeradorPlanilhaUsuario();
+					try {
+						geradorPlanilha.gerarPlanilhaUsuarios(usuario, caminhoEscolhido);
+						JOptionPane.showMessageDialog(null, "Planilha gerada com sucesso!", Constants.SUCESSO,
+								JOptionPane.INFORMATION_MESSAGE, null);
+
+					} catch (Exception ex) {
+						JOptionPane.showMessageDialog(null, "Erro ao gerar planilha!", "ATENÇÃO",
+								JOptionPane.ERROR_MESSAGE, null);
+						System.out.println("Erro ao gerar planilha de usuarios: " + ex.getMessage());
+					}
+				}
+			}
+		});
+		btnGerarXls.setFont(new Font("Tahoma", Font.BOLD, 14));
+		btnGerarXls.setIcon(new ImageIcon(PanelConsultaQuestoes.class.getResource("/imagens/excel.png")));
+		btnGerarXls.setEnabled(false);
+		
+		cbxTurno = new JComboBox<String>();
+		cbxTurno.setModel(new DefaultComboBoxModel
+				(new String[] {"SELECIONE O TURNO", TurnoEnum.MATUTINO.toString(), TurnoEnum.VESPERTINO.toString(), TurnoEnum.NOTURNO.toString()}));
+		
+		JLabel lblTurno = new JLabel("Turno");
+		
+		JLabel lblDe = new JLabel("de");
+		
+		lblTotalPaginas = new JLabel("");
+		
 		GroupLayout gl_panel = new GroupLayout(panel);
 		gl_panel.setHorizontalGroup(
 			gl_panel.createParallelGroup(Alignment.LEADING)
 				.addGroup(gl_panel.createSequentialGroup()
 					.addGap(272)
-					.addComponent(lblRelatorioDeUsuarios, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+					.addComponent(lblRelatorioDeUsuarios, GroupLayout.PREFERRED_SIZE, 273, Short.MAX_VALUE)
 					.addGap(253))
+				.addGroup(gl_panel.createSequentialGroup()
+					.addGap(250)
+					.addComponent(btnPaginaAnterior, GroupLayout.DEFAULT_SIZE, 93, Short.MAX_VALUE)
+					.addPreferredGap(ComponentPlacement.UNRELATED)
+					.addComponent(lblPaginaAtual, GroupLayout.PREFERRED_SIZE, 14, GroupLayout.PREFERRED_SIZE)
+					.addGap(2)
+					.addComponent(lblDe)
+					.addPreferredGap(ComponentPlacement.UNRELATED)
+					.addComponent(lblTotalPaginas, GroupLayout.PREFERRED_SIZE, 19, GroupLayout.PREFERRED_SIZE)
+					.addPreferredGap(ComponentPlacement.UNRELATED)
+					.addComponent(btnProximaPagina, GroupLayout.DEFAULT_SIZE, 86, Short.MAX_VALUE)
+					.addGap(249))
 				.addGroup(gl_panel.createSequentialGroup()
 					.addGap(71)
 					.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
-						.addComponent(tblListaDeUsuarios, Alignment.TRAILING, GroupLayout.DEFAULT_SIZE, 642, Short.MAX_VALUE)
 						.addGroup(gl_panel.createSequentialGroup()
-							.addComponent(lblNome)
-							.addGap(18)
+							.addComponent(btnGerarXls, GroupLayout.PREFERRED_SIZE, 161, GroupLayout.PREFERRED_SIZE)
+							.addContainerGap())
+						.addGroup(gl_panel.createSequentialGroup()
 							.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
+								.addComponent(tblListaDeUsuarios, Alignment.TRAILING, GroupLayout.DEFAULT_SIZE, 660, Short.MAX_VALUE)
 								.addGroup(gl_panel.createSequentialGroup()
-									.addComponent(btnLimpar, GroupLayout.DEFAULT_SIZE, 137, Short.MAX_VALUE)
-									.addPreferredGap(ComponentPlacement.UNRELATED)
-									.addComponent(btnExcluir, GroupLayout.DEFAULT_SIZE, 116, Short.MAX_VALUE)
-									.addPreferredGap(ComponentPlacement.RELATED))
-								.addComponent(txtNome, GroupLayout.DEFAULT_SIZE, 239, Short.MAX_VALUE))
-							.addGap(18)
-							.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
-								.addGroup(gl_panel.createSequentialGroup()
-									.addComponent(lblTipoDeUsuario)
-									.addPreferredGap(ComponentPlacement.UNRELATED)
-									.addComponent(cbxTipoUsuario, 0, 248, Short.MAX_VALUE))
-								.addGroup(Alignment.TRAILING, gl_panel.createSequentialGroup()
-									.addComponent(btnAlterar, GroupLayout.DEFAULT_SIZE, 146, Short.MAX_VALUE)
-									.addGap(18)
-									.addComponent(btnBuscar, GroupLayout.DEFAULT_SIZE, 148, Short.MAX_VALUE)))))
-					.addGap(67))
-				.addGroup(gl_panel.createSequentialGroup()
-					.addGap(264)
-					.addComponent(btnPaginaAnterior, GroupLayout.DEFAULT_SIZE, 92, Short.MAX_VALUE)
-					.addGap(18)
-					.addComponent(lblPaginaAtual)
-					.addGap(18)
-					.addComponent(btnProximaPagina, GroupLayout.DEFAULT_SIZE, 85, Short.MAX_VALUE)
-					.addGap(297))
+									.addComponent(lblNome)
+									.addPreferredGap(ComponentPlacement.RELATED)
+									.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
+										.addGroup(gl_panel.createSequentialGroup()
+											.addComponent(txtNome, GroupLayout.DEFAULT_SIZE, 139, Short.MAX_VALUE)
+											.addGap(18))
+										.addGroup(gl_panel.createSequentialGroup()
+											.addComponent(btnLimpar, GroupLayout.DEFAULT_SIZE, 121, Short.MAX_VALUE)
+											.addGap(36)))
+									.addGroup(gl_panel.createParallelGroup(Alignment.LEADING)
+										.addGroup(gl_panel.createSequentialGroup()
+											.addComponent(btnExcluir, GroupLayout.DEFAULT_SIZE, 117, Short.MAX_VALUE)
+											.addGap(86))
+										.addGroup(gl_panel.createSequentialGroup()
+											.addComponent(lblTurno)
+											.addPreferredGap(ComponentPlacement.RELATED)
+											.addComponent(cbxTurno, 0, 171, Short.MAX_VALUE)))
+									.addPreferredGap(ComponentPlacement.RELATED)
+									.addGroup(gl_panel.createParallelGroup(Alignment.TRAILING)
+										.addComponent(btnAlterar, GroupLayout.DEFAULT_SIZE, 109, Short.MAX_VALUE)
+										.addComponent(lblTipoDeUsuario))
+									.addPreferredGap(ComponentPlacement.RELATED)
+									.addGroup(gl_panel.createParallelGroup(Alignment.TRAILING)
+										.addComponent(cbxTipoUsuario, 0, 144, Short.MAX_VALUE)
+										.addGroup(gl_panel.createSequentialGroup()
+											.addGap(22)
+											.addComponent(btnBuscar, GroupLayout.DEFAULT_SIZE, 122, Short.MAX_VALUE)))))
+							.addGap(67))))
 		);
 		gl_panel.setVerticalGroup(
 			gl_panel.createParallelGroup(Alignment.LEADING)
@@ -225,8 +339,10 @@ public class PanelRelatorioDeUsuario extends JPanel {
 					.addGroup(gl_panel.createParallelGroup(Alignment.BASELINE)
 						.addComponent(lblNome)
 						.addComponent(txtNome, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-						.addComponent(lblTipoDeUsuario)
-						.addComponent(cbxTipoUsuario, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
+						.addComponent(lblTurno)
+						.addComponent(cbxTurno, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+						.addComponent(cbxTipoUsuario, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+						.addComponent(lblTipoDeUsuario))
 					.addGap(43)
 					.addGroup(gl_panel.createParallelGroup(Alignment.BASELINE)
 						.addComponent(btnLimpar)
@@ -239,15 +355,26 @@ public class PanelRelatorioDeUsuario extends JPanel {
 					.addGroup(gl_panel.createParallelGroup(Alignment.BASELINE)
 						.addComponent(btnPaginaAnterior)
 						.addComponent(lblPaginaAtual)
+						.addComponent(lblDe)
+						.addComponent(lblTotalPaginas)
 						.addComponent(btnProximaPagina))
-					.addContainerGap(86, Short.MAX_VALUE))
+					.addPreferredGap(ComponentPlacement.RELATED, GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+					.addComponent(btnGerarXls)
+					.addContainerGap())
 		);
 		panel.setLayout(gl_panel);
 		setLayout(groupLayout);
 
 	}
 
+	private void verificarBotoesPaginas() {
+		btnPaginaAnterior.setEnabled(paginaAtual > 1);
+		btnProximaPagina.setEnabled(paginaAtual < paginasTotal);
+	}
+
 	protected void buscarUsuariosSeletores() {
+		this.limparTabelaUsuario();
+		
 		lblPaginaAtual.setText(paginaAtual + "");
 		
 		RelatorioDeUsuarioSeletor relatorioUsuario = new RelatorioDeUsuarioSeletor();
@@ -268,17 +395,30 @@ public class PanelRelatorioDeUsuario extends JPanel {
 			}
 		}
 		
+		if (cbxTurno.getSelectedIndex() > 0) {
+			if (cbxTurno.getSelectedItem().equals("MATUTINO")) {
+				relatorioUsuario.setTurno(TurnoEnum.MATUTINO);
+			} else if (cbxTurno.getSelectedItem().equals("VESPERTINO")) {
+				relatorioUsuario.setTurno(TurnoEnum.VESPERTINO);
+			} else if (cbxTurno.getSelectedItem().equals("NOTURNO")) {
+				relatorioUsuario.setTurno(TurnoEnum.NOTURNO);
+			} else {
+				relatorioUsuario.setTurno(null);
+			}
+		}
+		
 		relatorioUsuario.setNome(txtNome.getText());
 		
 		usuario = usuarioController.relatorioUsuarioController(relatorioUsuario);
 		
+		paginasTotal = usuarioController.consultarTotalPaginas(relatorioUsuario);
+		lblTotalPaginas.setText(paginasTotal + "");
+		
+		verificarBotoesPaginas();
 		this.atualizarTabelaUsuario(usuario);
 	}
 
 	private void atualizarTabelaUsuario(List<UsuarioVO> usuario2) {
-		
-		this.limparTabelaUsuario();
-		
 		model = (DefaultTableModel) this.tblListaDeUsuarios.getModel();
 		
 		for(UsuarioVO usu: this.usuario) {
@@ -308,6 +448,8 @@ public class PanelRelatorioDeUsuario extends JPanel {
 			}
 		}
 		
+		habilitarBtnExcel();
+		
 //		for (UsuarioVO usuarioVO : usuario) {
 //			if(usuarioVO instanceof AlunoVO) {
 //				AlunoVO aluno = (AlunoVO) usuarioVO;
@@ -325,5 +467,15 @@ public class PanelRelatorioDeUsuario extends JPanel {
 
 	private void limparTabelaUsuario() {
 		tblListaDeUsuarios.setModel(new DefaultTableModel(new Object[][] {nomesColunas, }, nomesColunas));
+		btnPaginaAnterior.setEnabled(false);
+		btnProximaPagina.setEnabled(false);
+	}
+	
+	private void habilitarBtnExcel() {
+		if (tblListaDeUsuarios.getRowCount() > 1) {
+			btnGerarXls.setEnabled(true);
+		} else {
+			btnGerarXls.setEnabled(false);
+		}
 	}
 }
